requirements.txt
streamlit
google-genai
firebase-admin
google-cloud-firestore
python-dotenv
requests
